Fixes for action layer and paging support
The action layer was inconsistent in the formatting of the result,
with issue_comment_search returning a naked list of objects whilst
issue_search returned a dict with a count and a results key.  So
this commit fixes issue_comment_search.

Also adds basic pagination to the moderation controller, although
it is not yet rendered